ocr-invoice/requirements.txt
fastapi
uvicorn
puremagic
google-cloud-vision
google-genai
python-dotenv
psycopg2-binary
